Moved RefreshableObservableCollection to Collections directory.
