+++ src/Collections/RefreshableObservableCollection.cs
@@ -1,7 +1,7 @@
 using System.Collections.ObjectModel;
 using System.Collections.Specialized;
 
-namespace DLSS_Swapper.Datatypes
+namespace DLSS_Swapper.Collections
 {
     class RefreshableObservableCollection<T> : ObservableCollection<T>
     {
